// model/gcp/firebasecrashlytics/extensions/models/_lib/gcp.ts
export interface ExplicitGcpCredentials {
  accessToken?: string;
  credentialsJson?: string;
  project?: string;
}